refactor: [#34] Category 엔티티와 DishRef 엔티티 연관관계 매핑
